refactor: cleanup AccountService methods and tuple + put back /accounts/id/info endpoint
