+++ store/src/main/java/com/zextras/mailbox/api/rest/resource/AccountResource.java
@@ -6,23 +6,18 @@
 
 package com.zextras.mailbox.api.rest.resource;
 
+import com.zextras.mailbox.api.rest.resource.dto.AccountInfoResponse;
 import com.zextras.mailbox.api.rest.response.ErrorResponse;
 import com.zextras.mailbox.api.rest.service.AccountService;
-import com.zimbra.common.account.ZAttrProvisioning;
-import com.zimbra.common.account.ZAttrProvisioning.AccountStatus;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.cs.account.Account;
-import com.zimbra.cs.account.AuthToken;
 import com.zimbra.cs.account.AuthTokenException;
-import com.zimbra.cs.account.Domain;
-import com.zimbra.cs.account.Provisioning;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import java.util.List;
-import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
 import javax.ws.rs.Consumes;
@@ -84,7 +79,7 @@ public class AccountResource {
 			content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
 	@ApiResponse(responseCode = "500", description = "Internal server error",
 			content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-	@Path("/{id}")
+	@Path("/{id}/info")
 	public Response getAccountInfo(
 			@Parameter(description = "The account ID") @PathParam("id") String id) {
 		return accountService.getAccount(id)
@@ -116,8 +111,8 @@ public class AccountResource {
 					.entity(new ErrorResponse("Missing auth token"))
 					.build();
 		}
-		return accountService.getAccountAndAuthToken(token)
-				.map(tuple -> Response.ok(AccountInfoResponse.from(tuple._1(), tuple._2())).build())
+		return accountService.getAuthToken(token)
+				.mapTry(authToken -> Response.ok(AccountInfoResponse.from(authToken)).build())
 				.recover(e -> switch (e) {
 					case AuthTokenException ignored -> Response.status(Response.Status.UNAUTHORIZED)
 							.entity(new ErrorResponse("Invalid auth token"))
@@ -207,89 +202,4 @@ public class AccountResource {
 
 	public record BatchRequest(List<String> ids, List<String> emails) {}
 
-	public record CarbonioFeature(String name, Object value) {
-		public static CarbonioFeature map(Entry<String, Object> entry) {
-			return new CarbonioFeature(entry.getKey(), entry.getValue());
-		}
-	}
-
-	public record AccountInfoResponse(String id, String name, String displayName, String cosId,
-																		String domainId, String domain, AccountStatus status, boolean isGlobalAdmin,
-																		boolean isExternal, String locale, Map<String, Boolean> features,
-																		Map<String, String> capabilities, Long sessionLifetimeMs) {
-
-		/**
-		 * Returns the public service URL for the account's domain in the same format as the
-		 * legacy SOAP {@code GetAccountInfoResponse.getPublicURL()} field:
-		 * {@code {zimbraPublicServiceProtocol}://{zimbraPublicServiceHostname}}.
-		 * Falls back to the plain domain name if the public service hostname is not configured.
-		 */
-		private static String getPublicServiceUrl(Account account) {
-			try {
-				Domain domain = Provisioning.getInstance().getDomain(account);
-				if (domain != null) {
-					String hostname = domain.getAttr(Provisioning.A_zimbraPublicServiceHostname, null);
-					if (hostname != null && !hostname.isBlank()) {
-						String proto = domain.getAttr(Provisioning.A_zimbraPublicServiceProtocol, "http");
-						return proto + "://" + hostname;
-					}
-				}
-			} catch (ServiceException ignored) {
-				// fall through to domain name fallback
-			}
-			return account.getDomainName();
-		}
-
-		public static AccountInfoResponse from(Account account) {
-			var features = account.getAttrs().entrySet().stream()
-					.filter(entry -> entry.getKey().startsWith("carbonioFeature"))
-					.collect(Collectors.toMap(Entry::getKey, entry -> Boolean.parseBoolean(entry.getValue().toString())));
-			var capabilities = account.getAttrs().entrySet().stream()
-					.filter(entry -> {
-						String key = entry.getKey();
-						return key.startsWith("carbonioWsc") || key.startsWith("carbonioFiles")
-								|| key.startsWith("carbonioTasks") || key.startsWith("carbonioDocs")
-								|| key.startsWith("carbonioPreview");
-					})
-					.collect(Collectors.toMap(Entry::getKey, entry -> entry.getValue().toString()));
-			boolean isExternal;
-			try {
-				isExternal = account.isAccountExternal();
-			} catch (ServiceException e) {
-				// If we cannot determine whether the account is external, default to true.
-				// An internal account failing this check implies a non-standard transport
-				// configuration, which is more consistent with an external account.
-				isExternal = true;
-			}
-			return new AccountInfoResponse(account.getId(), account.getName(), account.getDisplayName(),
-					account.getCOSId(), account.getDomainId(), getPublicServiceUrl(account),
-					account.getAccountStatus(), account.isIsAdminAccount(), isExternal,
-					account.getLocaleAsString(), features, capabilities, null);
-		}
-
-		public static AccountInfoResponse from(Account account, AuthToken authToken) {
-			var features = account.getAttrs().entrySet().stream()
-					.filter(entry -> entry.getKey().startsWith("carbonioFeature"))
-					.collect(Collectors.toMap(Entry::getKey, entry -> Boolean.parseBoolean(entry.getValue().toString())));
-			var capabilities = account.getAttrs().entrySet().stream()
-					.filter(entry -> {
-						String key = entry.getKey();
-						return key.startsWith("carbonioWsc") || key.startsWith("carbonioFiles")
-								|| key.startsWith("carbonioTasks") || key.startsWith("carbonioDocs")
-								|| key.startsWith("carbonioPreview");
-					})
-					.collect(Collectors.toMap(Entry::getKey, entry -> entry.getValue().toString()));
-			boolean isExternal;
-			try {
-				isExternal = account.isAccountExternal();
-			} catch (ServiceException e) {
-				isExternal = true;
-			}
-			long sessionLifetimeMs = authToken.getExpires() - System.currentTimeMillis();
-			return new AccountInfoResponse(account.getId(), account.getName(), account.getDisplayName(),
-					account.getCOSId(), account.getDomainId(), getPublicServiceUrl(account),
-					account.getAccountStatus(), account.isIsAdminAccount(), isExternal,
-					account.getLocaleAsString(), features, capabilities, sessionLifetimeMs);
-		}
-	}
 }
